bench: generate csharp category influence workload
