+++ examples/benchmark/generate_pipeline.py
@@ -1622,6 +1622,206 @@ fn main() {{
 '''
 
 
+def generate_csharp_query_category_influence(article_cats, category_parents, root_cats, n=5, max_depth=10):
+    roots = sorted(root_cats) if root_cats else ["Physics"]
+    root_entries = ", ".join(f'"{root}"' for root in roots)
+
+    return f'''// Category influence benchmark (C# query engine)
+using System;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.Globalization;
+using System.IO;
+using System.Linq;
+using UnifyWeaver.QueryRuntime;
+
+class Program
+{{
+    const double N = {float(n)};
+    const int MAX_DEPTH = {max_depth};
+    static readonly HashSet<string> ROOTS = new(new[] {{ {root_entries} }}, StringComparer.Ordinal);
+
+    static void Main(string[] args)
+    {{
+        if (args.Length < 2)
+        {{
+            Console.Error.WriteLine("Usage: program <category_parent.tsv> <article_category.tsv>");
+            Environment.Exit(1);
+        }}
+
+        var swTotal = Stopwatch.StartNew();
+        var swLoad = Stopwatch.StartNew();
+
+        var provider = new InMemoryRelationProvider();
+        var edgeId = new PredicateId("category_parent", 2);
+        var predId = new PredicateId("category_ancestor", 3);
+        var articleCategories = new Dictionary<string, List<string>>(StringComparer.Ordinal);
+
+        foreach (var (line, i) in File.ReadLines(args[0]).Select((l, i) => (l, i)))
+        {{
+            if (i == 0 && (line.StartsWith("child") || line.StartsWith("article")))
+            {{
+                continue;
+            }}
+
+            var parts = line.Split('\\t', 2);
+            if (parts.Length == 2)
+            {{
+                provider.AddFact(edgeId, parts[0], parts[1]);
+            }}
+        }}
+
+        foreach (var (line, i) in File.ReadLines(args[1]).Select((l, i) => (l, i)))
+        {{
+            if (i == 0 && (line.StartsWith("article") || line.StartsWith("child")))
+            {{
+                continue;
+            }}
+
+            var parts = line.Split('\\t', 2);
+            if (parts.Length != 2)
+            {{
+                continue;
+            }}
+
+            if (!articleCategories.TryGetValue(parts[0], out var categories))
+            {{
+                categories = new List<string>();
+                articleCategories[parts[0]] = categories;
+            }}
+
+            categories.Add(parts[1]);
+        }}
+
+        swLoad.Stop();
+
+        var plan = new QueryPlan(
+            predId,
+            new PathAwareTransitiveClosureNode(edgeId, predId, 1, 1, MAX_DEPTH, TableMode.All),
+            true,
+            new int[] {{ 0 }}
+        );
+
+        var uniqueSeeds = new HashSet<string>(StringComparer.Ordinal);
+        foreach (var categories in articleCategories.Values)
+        {{
+            foreach (var category in categories)
+            {{
+                uniqueSeeds.Add(category);
+            }}
+        }}
+
+        var seedParams = uniqueSeeds
+            .OrderBy(category => category, StringComparer.Ordinal)
+            .Select(category => new object[] {{ category }})
+            .ToList();
+
+        var swQuery = Stopwatch.StartNew();
+        var executor = new QueryExecutor(provider, new QueryExecutorOptions(ReuseCaches: false));
+        var rows = executor.Execute(plan, seedParams).ToList();
+        swQuery.Stop();
+
+        var swAgg = Stopwatch.StartNew();
+        var ancestorIndex = new Dictionary<string, List<(string Ancestor, int Hops)>>(StringComparer.Ordinal);
+        foreach (var row in rows)
+        {{
+            var source = row[0]?.ToString() ?? "";
+            var ancestor = row[1]?.ToString() ?? "";
+            var hops = Convert.ToInt32(row[2], CultureInfo.InvariantCulture);
+            if (!ancestorIndex.TryGetValue(source, out var bucket))
+            {{
+                bucket = new List<(string, int)>();
+                ancestorIndex[source] = bucket;
+            }}
+
+            bucket.Add((ancestor, hops));
+        }}
+
+        var rootId = new PredicateId("root_category", 1);
+        var weightId = new PredicateId("article_root_weight", 3);
+        var resultId = new PredicateId("category_influence", 2);
+
+        foreach (var root in ROOTS.OrderBy(x => x, StringComparer.Ordinal))
+        {{
+            provider.AddFact(rootId, root);
+        }}
+
+        foreach (var (_, categories) in articleCategories.OrderBy(kvp => kvp.Key, StringComparer.Ordinal))
+        {{
+            foreach (var category in categories)
+            {{
+                if (ROOTS.Contains(category))
+                {{
+                    provider.AddFact(weightId, "_", category, 1.0);
+                }}
+
+                if (!ancestorIndex.TryGetValue(category, out var ancestors))
+                {{
+                    continue;
+                }}
+
+                foreach (var (ancestor, hops) in ancestors)
+                {{
+                    if (!ROOTS.Contains(ancestor))
+                    {{
+                        continue;
+                    }}
+
+                    var distance = hops + 1;
+                    var weight = Math.Pow(distance, -N);
+                    provider.AddFact(weightId, "_", ancestor, weight);
+                }}
+            }}
+        }}
+
+        var aggregatePlan = new QueryPlan(
+            resultId,
+            new ProjectionNode(
+                new SelectionNode(
+                    new AggregateNode(
+                        new RelationScanNode(rootId),
+                        weightId,
+                        AggregateOperation.Sum,
+                        tuple => new object[] {{ Wildcard.Value, tuple[0], Wildcard.Value }},
+                        Array.Empty<int>(),
+                        2,
+                        2),
+                    tuple => QueryExecutor.CompareValues(tuple[1], 0) > 0),
+                tuple => new object[] {{ tuple[0], tuple[1] }}),
+            false,
+            null
+        );
+
+        var aggregateRows = executor.Execute(aggregatePlan).ToList();
+        swAgg.Stop();
+        swTotal.Stop();
+
+        var results = aggregateRows
+            .Select(row => (
+                Root: row[0]?.ToString() ?? "",
+                Score: Convert.ToDouble(row[1], CultureInfo.InvariantCulture)))
+            .OrderByDescending(item => item.Score)
+            .ThenBy(item => item.Root, StringComparer.Ordinal)
+            .ToList();
+
+        Console.WriteLine("root_category\\tinfluence_score");
+        foreach (var item in results)
+        {{
+            Console.WriteLine($"{{item.Root}}\\t{{item.Score.ToString("F12", CultureInfo.InvariantCulture)}}");
+        }}
+
+        Console.Error.WriteLine($"load_ms={{swLoad.ElapsedMilliseconds}}");
+        Console.Error.WriteLine($"query_ms={{swQuery.ElapsedMilliseconds}}");
+        Console.Error.WriteLine($"aggregation_ms={{swAgg.ElapsedMilliseconds}}");
+        Console.Error.WriteLine($"total_ms={{swTotal.ElapsedMilliseconds}}");
+        Console.Error.WriteLine($"seed_count={{seedParams.Count}}");
+        Console.Error.WriteLine($"tuple_count={{rows.Count}}");
+        Console.Error.WriteLine($"root_count={{results.Count}}");
+    }}
+}}
+'''
+
+
 def generate_csharp_weighted_shortest_path(article_cats, category_parents, root_cats, n=5, max_depth=10):
     root = list(root_cats)[0] if root_cats else "Physics"
 
@@ -1758,6 +1958,7 @@ GENERATORS = {
         'csharp': generate_csharp_weighted_shortest_path,
     },
     'category_influence': {
+        'csharp_query': generate_csharp_query_category_influence,
         'go': generate_go_category_influence,
         'rust': generate_rust_category_influence,
     },
